Add files via upload
Exercise 25.17.(Data compression: Huffman coding animation) Write a program that enables 
the user to enter text and displays the Huffman coding tree based on the text, as 
shown in Figure 25.23a. Display the weight of the subtree inside the subtree’s 
root circle. Display each leaf node’s character. Display the encoded bits for 
the text in a label. When the user clicks the Decode Text button, a bit string is 
decoded into text displayed in the label, as shown in Figure 25.23b.